Feature: PIN-3587 - Fix AutocompleteSingle queries  (#486)
Before this, if a value was selected in the input field, the suggestion menu would be empty, as if no other options were possible. This fixes it, and now when the input field has a selected value, the suggestion menu shows a default query.

The weird behavior is due to the fact that filtering results happens by querying (with the q parameter) the backend service. When a value is selected, only the selected value is returned from the backend. 

This is a temporary fix, will need to agree a more scalable fix with the backend team